Use single @output_buffer when rendering
When rendering components we sometimes run into incompatibilities with
Rails due to Rails expecting a single `@output_buffer` when rendering
templates. This isn't true when using view components due to each
component having its own `@output_buffer`.

The most obvious issue I've run into due us not having a single
`@output_buffer` can be recreated by repeating the following:

1. Create a form using `form_for`
2. Pass the `f` variable to a component
3. Call `f.label` while passing it a block

This will cause the block content to be rendered incorrectly, usually
above the resulting `<label>` tag.

The tl;dr of that behavior is that `form_for` captures the
`view_context` of the template it's called in. When trying to evaluate
the block passed to `f.label` it calls `view_context.capture` *but* the
block was defined and is run in the context of the component, not
`view_context`! This causes the block content to be rendered in the
component's `@output_buffer` (since it's not capturing).

This attempts to resolve the issue by tracking and using a single
`@output_buffer`. When a component is rendered we find the top-level
`ActionView::Base` instance in the render hierarchy and use its
`@output_buffer`. We then tell the top-level renderer that we are one of
its children.

Now, when *any* `ActionView::Base` instance in the rendering hierarchy
changes its `@output_buffer`, it propagates that change to all
`ActionView::Base` instances. This matches Rails' expectations and
results in the previously broken behavior working as-expected.